requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
scikit-surprise==1.1.3
pillow==10.0.0
ttkthemes==3.2.2
colorama==0.4.6
tqdm==4.66.1
matplotlib==3.7.1
seaborn==0.12.2
